move useResizeObserver to helper functions and delete old individual line graph component
